Use Compose 1.7.0 final version and merge taskListScreen impls for desktop & Android using navigation suite
